Added pdwBuilder.py, cleaned up a few erroneous function calls and exception handling clauses. error.py: PEP8 instruments.py: Cleaned up imports and exception handling. Removed PDW building functions. pdwBuilder.py: Moved PDW building functions here. wfmBuilder.py: Cleaned up imports, fixed some plotting and directory finding functions. setup.py: Fixed formatting.
